update: move models folder to package/app
